auth_api/python/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
boto3==1.26.62
botocore==1.29.62
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Django==4.0.6
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.0
MarkupSafe==2.1.2
marshmallow==3.19.0
mysql-connector-python==8.0.32
numpy==1.23.0
packaging==23.0
pandas==1.4.3
protobuf==3.20.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.2
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.10
Werkzeug==2.2.2